:dress: rename package name
