+++ pkg/pointer/pointer.go
